Correct data type of series save_no field (#99)

+++ Sma5h/ResourceProviders/Sma5h.ResourceProviders.PrcExt/Data/Ui/Param/Database/PrcUiSeriesDatabase.cs
@@ -28,7 +28,7 @@ namespace Sma5h.Data.Ui.Param.Database
             public sbyte DispOrderSound { get; set; }
 
             [PrcHexMapping("save_no")]
-            public short SaveNo { get; set; }
+            public sbyte SaveNo { get; set; }
 
             [PrcHexMapping(0x1c38302364)]
             public bool Unk1 { get; set; }
